refactor(worktree): split WorktreeHeader into focused subcomponents
- Extract useGitHubBadgeTooltip shared hook (eliminates ~52 lines of duplicated Issue/PR badge lifecycle boilerplate)
- Extract UpstreamSyncBadge, IssueBadge, PRBadge, EnvironmentPopover, CollapsedSessionIndicators
- Extract WorktreeActionsToolbar, MainWorktreeSecondaryRow, NonMainSecondaryRow
- WorktreeHeader.tsx shrinks from 1059→~270 lines as orchestrator
- Public API (WorktreeHeaderProps, WorktreeHeader export) unchanged
- All 78 existing tests pass with zero expectation changes
